requirements.txt
yfinance
matplotlib
python-telegram-bot
pycoingecko
requests
